Added additional paths for the null trajectory
Essentially if you are on the left or right side it will do a partial path of going for its side switch if we trust the centre robot. If we don't trust the center robot (indicated by a chooser on the smartdashboard) then it will try to go for the switch.