Updated path selection logic
Primary roll is drive team choice (switch for RP, scale when control matters to alliance). If primary is inaccessible or cannot be determined due to a field error, code will attempt a secondary solution; if secondary is inaccessible or cannot be determined due to a field error, then a third solution of just moving to right or left side is considered. If there is no solution the robot remains stationary.